move: application page out of components
