+++ apps/admin-ui/src/spa/applications/TimeSelector.tsx
