+++ apps/admin-ui/src/spa/applications/queries.tsx
